Add Content attributes
