+++ README.md
@@ -1,5 +1,7 @@
 # CakeContent plugin for CakePHP
 
+Content management plugin
+
 ## Installation
 
 You can install this plugin into your CakePHP application using [composer](http://getcomposer.org).
